Major refactor to more cleanly separate functionality.
