+++ include/util.h
@@ -29,87 +29,4 @@
  */
 void *secure_memset(void *v, int c, size_t n);
 
-/**
- * <pre>
- * This function creates a new CharBuf that contains the file extension of a file name sting in a CharBuf
- * </pre>
- *
- * @param[in] buf The CharBuf that contains the file name string
- * @param[out] ext The integer representation of the file extension type
- *
- * @return 0 on success, 1 on error
- */
-int get_file_ext(CharBuf buf, int *ext);
-
-/**
- * <pre>
- * Load key from location stated by Key ID
- * <pre>
- *
- * @param[in] key_values.key_id The Key Identifier
- * @param[in] key_values.key_id_len Length of Key Identifier
- * @param[out] key_values.key The key value
- * @param[out] key_values.key_len The length of the key
- *
- * @return 0 on success, 1 on error
- */
-int key_load(KeyEntry * key_values);
-
-/**
- * <pre>
- * URI parsing of Key ID per RFC 8089 (The "file" URI Scheme) and RFC 959 (FILE TRANSFER PROTOCOL (FTP)) with
- * RFC 1738 (Uniform Resource Locators) Section 3.1 (Common Internet Scheme Syntax).
- * File path will be assumed to be the absolute path to file for production but uses local path for testing.
- * <pre>
- *
- * @param[in] key_id The Key Identifier to be parsed
- * @param[out] uri.type The type of scheme used for key_id
- * @param[out] uri.auth The file authority
- * @param[out] uri.path The path to file that contains the key
- * @param[out] uri.file The file that contains the key
- * @param[out] uri.host The IP address for the key host server
- * @param[out] uri.port The port to access the key host server
- * @param[out] uri.url_path The location path on key host server
- *
- * @return 0 on success, 1 on error
- */
-int key_id_parse(CharBuf key_id, URIValues * uri);
-
-/**
- * <pre>
- * Using key_id to check if there is actual file
- * <pre>
- *
- * @param[in] key_id The Identifier for the Key which is also file path and name
- *
- * @return 0 on success, 1 on error
- */
-int file_check(char *file_path);
-
-/**
- * <pre>
- * Check if the file path provided is valid.
- * </pre>
- *
- * @param[in] file_path The file path provided to be checked
- *
- * @return 0 if success, 1 if error
- */
-
-int encodeBase64Data(unsigned char *raw_data, size_t raw_data_size, unsigned char **base64_data, size_t * base64_data_size);
-
-/**
- * <pre>
- * Reads an unsigned string encoded in base64 encoding and produces a string of raw bytes
- * </pre>
- *
- * @param[in] base64_data The data to be decoded
- * @param[in] base64data The size of the data to be decoded
- * @param[out] raw_data The raw bytes retrieved from base64_data
- * @param[out] raw_data_size The number of raw bytes
- *
- * @return 0 if success, 1 if error
- */
-int decodeBase64Data(unsigned char *base64_data, size_t b64_data_size, unsigned char **raw_data, size_t * raw_data_size);
-
 #endif /* INCLUDE_UTIL_H_ */
