+++ src/util/key_table.c
@@ -4,51 +4,46 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#include <stdbool.h>
 #include <string.h>
-#include <pthread.h>
 
+#include "pelz_io.h"
 #include <key_table.h>
 #include <util.h>
 #include <pelz_request_handler.h>
 #include <CharBuf.h>
 #include <pelz_log.h>
 
+KeyTable key_table;
+
 //Initialize key table
-int key_table_init(KeyTable * key_table)
+int key_table_init(void)
 {
-  if ((key_table->entries = (KeyEntry *) malloc(sizeof(KeyEntry))) == NULL)
+  if ((key_table.entries = (KeyEntry *) malloc(sizeof(KeyEntry))) == NULL)
   {
     pelz_log(LOG_ERR, "Key List Space Allocation Error");
     return (1);
   }
 
-  key_table->num_entries = 0;
-  key_table->mem_size = 0;
-  if (pthread_mutex_init(&key_table->lock, NULL) != 0)
-  {
-    pelz_log(LOG_ERR, "Mutex Init has failed.");
-    return (1);
-  }
+  key_table.num_entries = 0;
+  key_table.mem_size = 0;
   return (0);
 }
 
 //Destroy key table
-int key_table_destroy(KeyTable * key_table)
+int key_table_destroy(void)
 {
   pelz_log(LOG_DEBUG, "Key Table Destroy Function Starting");
-  pthread_mutex_lock(&key_table->lock);
-  if (key_table->num_entries >= 0)
+  if (key_table.num_entries >= 0)
   {
-    for (int i = 0; i < key_table->num_entries; i++)
+    for (int i = 0; i < key_table.num_entries; i++)
     {
-      if (key_table->entries[i].key_id.len != 0)
+      if (key_table.entries[i].key_id.len != 0)
       {
-        freeCharBuf(&key_table->entries[i].key_id);
+        freeCharBuf(&key_table.entries[i].key_id);
       }
-      if (key_table->entries[i].key.len != 0)
+      if (key_table.entries[i].key.len != 0)
       {
-        secureFreeCharBuf(&key_table->entries[i].key);
+        secureFreeCharBuf(&key_table.entries[i].key);
       }
     }
   }
@@ -58,29 +53,26 @@ int key_table_destroy(KeyTable * key_table)
     return (1);
   }
 
-  pthread_mutex_unlock(&key_table->lock);
-  pthread_mutex_destroy(&key_table->lock);
   //Free the storage allocated for the hash table
-  free(key_table->entries);
+  free(key_table.entries);
   pelz_log(LOG_DEBUG, "Key Table Destroy Function Complete");
   return (0);
 }
 
-int key_table_delete(CharBuf key_id, KeyTable * key_table)
+int key_table_delete(CharBuf key_id)
 {
   int index;
 
   index = 0;
-  pthread_mutex_lock(&key_table->lock);
-  for (int i = 0; i < key_table->num_entries; i++)
+  for (int i = 0; i < key_table.num_entries; i++)
   {
-    if (cmpCharBuf(key_id, key_table->entries[i].key_id) == 0)
+    if (cmpCharBuf(key_id, key_table.entries[i].key_id) == 0)
     {
-      key_table->mem_size = key_table->mem_size -
-        ((key_table->entries[i].key.len * sizeof(char)) + (key_table->entries[i].key_id.len * sizeof(char)) +
+      key_table.mem_size = key_table.mem_size -
+        ((key_table.entries[i].key.len * sizeof(char)) + (key_table.entries[i].key_id.len * sizeof(char)) +
         (2 * sizeof(size_t)));
-      freeCharBuf(&key_table->entries[i].key_id);
-      secureFreeCharBuf(&key_table->entries[i].key);
+      freeCharBuf(&key_table.entries[i].key_id);
+      secureFreeCharBuf(&key_table.entries[i].key);
       index = i + 1;
       break;
     }
@@ -88,28 +80,25 @@ int key_table_delete(CharBuf key_id, KeyTable * key_table)
   if (index == 0)
   {
     pelz_log(LOG_ERR, "Key ID not found.");
-    pthread_mutex_unlock(&key_table->lock);
     return (1);
   }
   else
   {
-    for (int i = index; i < key_table->num_entries; i++)
+    for (int i = index; i < key_table.num_entries; i++)
     {
-      key_table->entries[i - 1] = key_table->entries[i];
+      key_table.entries[i - 1] = key_table.entries[i];
     }
-    key_table->num_entries -= 1;
-    if ((key_table->entries = (KeyEntry *) realloc(key_table->entries, (key_table->num_entries) * sizeof(KeyEntry))) == NULL)
+    key_table.num_entries -= 1;
+    if ((key_table.entries = (KeyEntry *) realloc(key_table.entries, (key_table.num_entries) * sizeof(KeyEntry))) == NULL)
     {
       pelz_log(LOG_ERR, "Key List Space Reallocation Error");
-      pthread_mutex_unlock(&key_table->lock);
       return (1);
     }
   }
-  pthread_mutex_unlock(&key_table->lock);
   return (0);
 }
 
-int key_table_add(CharBuf key_id, CharBuf * key, KeyTable * key_table)
+int key_table_add(CharBuf key_id, CharBuf * key)
 {
   KeyEntry tmp_entry;
   size_t max_mem_size;
@@ -117,7 +106,7 @@ int key_table_add(CharBuf key_id, CharBuf * key, KeyTable * key_table)
 
   max_mem_size = 1000000;
 
-  if (key_table->mem_size >= max_mem_size)
+  if (key_table.mem_size >= max_mem_size)
   {
     pelz_log(LOG_ERR, "Key Table memory allocation greater then specified limit.");
     return (1);
@@ -133,8 +122,7 @@ int key_table_add(CharBuf key_id, CharBuf * key, KeyTable * key_table)
     return (1);
   }
 
-  pthread_mutex_lock(&key_table->lock);
-  if (!key_table_lookup(tmp_entry.key_id, &tmpkey, key_table, true))
+  if (!key_table_lookup(tmp_entry.key_id, &tmpkey))
   {
     if (cmpCharBuf(tmpkey, tmp_entry.key) == 0)
     {
@@ -143,7 +131,6 @@ int key_table_add(CharBuf key_id, CharBuf * key, KeyTable * key_table)
       freeCharBuf(&tmp_entry.key_id);
       secureFreeCharBuf(&tmp_entry.key);
       secureFreeCharBuf(&tmpkey);
-      pthread_mutex_unlock(&key_table->lock);
       return (0);
     }
     else
@@ -152,51 +139,36 @@ int key_table_add(CharBuf key_id, CharBuf * key, KeyTable * key_table)
       freeCharBuf(&tmp_entry.key_id);
       secureFreeCharBuf(&tmp_entry.key);
       secureFreeCharBuf(&tmpkey);
-      pthread_mutex_unlock(&key_table->lock);
       return (1);
     }
   }
 
-  if ((key_table->entries = (KeyEntry *) realloc(key_table->entries, (key_table->num_entries + 1) * sizeof(KeyEntry))) == NULL)
+  if ((key_table.entries = (KeyEntry *) realloc(key_table.entries, (key_table.num_entries + 1) * sizeof(KeyEntry))) == NULL)
   {
     pelz_log(LOG_ERR, "Key List Space Reallocation Error");
     freeCharBuf(&tmp_entry.key_id);
     secureFreeCharBuf(&tmp_entry.key);
-    pthread_mutex_unlock(&key_table->lock);
     return (1);
   }
-  key_table->entries[key_table->num_entries] = tmp_entry;
-  key_table->num_entries++;
-  key_table->mem_size =
-    key_table->mem_size + ((tmp_entry.key.len * sizeof(char)) + (tmp_entry.key_id.len * sizeof(char)) + (2 * sizeof(size_t)));
+  key_table.entries[key_table.num_entries] = tmp_entry;
+  key_table.num_entries++;
+  key_table.mem_size =
+    key_table.mem_size + ((tmp_entry.key.len * sizeof(char)) + (tmp_entry.key_id.len * sizeof(char)) + (2 * sizeof(size_t)));
   pelz_log(LOG_INFO, "Key Added");
-  pthread_mutex_unlock(&key_table->lock);
   *key = copyBytesFromBuf(tmp_entry.key, 0);
   return (0);
 }
 
-int key_table_lookup(CharBuf key_id, CharBuf * key, KeyTable * key_table, bool hasLock)
+int key_table_lookup(CharBuf key_id, CharBuf * key)
 {
-  if (!hasLock)
+  for (int i = 0; i < key_table.num_entries; i++)
   {
-    pthread_mutex_lock(&key_table->lock);
-  }
-  for (int i = 0; i < key_table->num_entries; i++)
-  {
-    if (cmpCharBuf(key_id, key_table->entries[i].key_id) == 0)
+    if (cmpCharBuf(key_id, key_table.entries[i].key_id) == 0)
     {
-      *key = newCharBuf(key_table->entries[i].key.len);
-      memcpy(key->chars, key_table->entries[i].key.chars, key->len);
-      if (!hasLock)
-      {
-        pthread_mutex_unlock(&key_table->lock);
-      }
+      *key = newCharBuf(key_table.entries[i].key.len);
+      memcpy(key->chars, key_table.entries[i].key.chars, key->len);
       return (0);
     }
-  }
-  if (!hasLock)
-  {
-    pthread_mutex_unlock(&key_table->lock);
   }
   return (1);
 }
