+++ src/util/pelz_json_parser.c
@@ -12,7 +12,7 @@
 #include <CharBuf.h>
 #include <pelz_log.h>
 
-int request_decoder(CharBuf request, RequestValues * request_values)
+int request_decoder(CharBuf request, RequestType * request_type, CharBuf * key_id, CharBuf * data)
 {
   cJSON *json;
   char *str = NULL;
@@ -33,11 +33,11 @@ int request_decoder(CharBuf request, RequestValues * request_values)
     cJSON_Delete(json);
     return (1);
   }
-  request_values->request_type = cJSON_GetObjectItemCaseSensitive(json, "request_type")->valueint;
-  switch (request_values->request_type)
+  *request_type = cJSON_GetObjectItemCaseSensitive(json, "request_type")->valueint;
+  switch (*request_type)
   {
   case REQ_ENC:
-    if (encrypt_parser(json, request_values))
+    if (encrypt_parser(json, key_id, data))
     {
       pelz_log(LOG_ERR, "Encrypt Request Parser Error");
       cJSON_Delete(json);
@@ -45,7 +45,7 @@ int request_decoder(CharBuf request, RequestValues * request_values)
     }
     break;
   case REQ_DEC:
-    if (decrypt_parser(json, request_values))
+    if (decrypt_parser(json, key_id, data))
     {
       pelz_log(LOG_ERR, "Decrypt Request Parser Error");
       cJSON_Delete(json);
@@ -81,37 +81,37 @@ int error_message_encoder(CharBuf * message, char *err_message)
   return (0);
 }
 
-int message_encoder(RequestValues request_values, CharBuf * message)
+int message_encoder(RequestType request_type, CharBuf key_id, CharBuf data, CharBuf * message)
 {
   cJSON *root;
   char *tmp = NULL;
 
   root = cJSON_CreateObject();
-  switch (request_values.request_type)
+  switch (request_type)
   {
   case REQ_ENC:
-    tmp = calloc((request_values.key_id.len + 1), sizeof(char));
-    memcpy(tmp, request_values.key_id.chars, request_values.key_id.len);
+    tmp = calloc((key_id.len + 1), sizeof(char));
+    memcpy(tmp, key_id.chars, key_id.len);
     cJSON_AddItemToObject(root, "key_id", cJSON_CreateString(tmp));
     free(tmp);
-    cJSON_AddItemToObject(root, "key_id_len", cJSON_CreateNumber(request_values.key_id.len));
-    tmp = calloc((request_values.data_out.len + 1), sizeof(char));
-    memcpy(tmp, request_values.data_out.chars, request_values.data_out.len);
+    cJSON_AddItemToObject(root, "key_id_len", cJSON_CreateNumber(key_id.len));
+    tmp = calloc((data.len + 1), sizeof(char));
+    memcpy(tmp, data.chars, data.len);
     cJSON_AddItemToObject(root, "enc_out", cJSON_CreateString(tmp));
     free(tmp);
-    cJSON_AddItemToObject(root, "enc_out_len", cJSON_CreateNumber(request_values.data_out.len));
+    cJSON_AddItemToObject(root, "enc_out_len", cJSON_CreateNumber(data.len));
     break;
   case REQ_DEC:
-    tmp = calloc((request_values.key_id.len + 1), sizeof(char));
-    memcpy(tmp, request_values.key_id.chars, request_values.key_id.len);
+    tmp = calloc((key_id.len + 1), sizeof(char));
+    memcpy(tmp, key_id.chars, key_id.len);
     cJSON_AddItemToObject(root, "key_id", cJSON_CreateString(tmp));
     free(tmp);
-    cJSON_AddItemToObject(root, "key_id_len", cJSON_CreateNumber(request_values.key_id.len));
-    tmp = calloc((request_values.data_out.len + 1), sizeof(char));
-    memcpy(tmp, request_values.data_out.chars, request_values.data_out.len);
+    cJSON_AddItemToObject(root, "key_id_len", cJSON_CreateNumber(key_id.len));
+    tmp = calloc((data.len + 1), sizeof(char));
+    memcpy(tmp, data.chars, data.len);
     cJSON_AddItemToObject(root, "dec_out", cJSON_CreateString(tmp));
     free(tmp);
-    cJSON_AddItemToObject(root, "dec_out_len", cJSON_CreateNumber(request_values.data_out.len));
+    cJSON_AddItemToObject(root, "dec_out_len", cJSON_CreateNumber(data.len));
     break;
   default:
     pelz_log(LOG_ERR, "Request Type not recognized.");
@@ -132,7 +132,7 @@ int message_encoder(RequestValues request_values, CharBuf * message)
   return (0);
 }
 
-int encrypt_parser(cJSON * json, RequestValues * request_values)
+int encrypt_parser(cJSON * json, CharBuf * key_id, CharBuf * data)
 {
   if (!cJSON_HasObjectItem(json, "key_id"))
   {
@@ -159,67 +159,65 @@ int encrypt_parser(cJSON * json, RequestValues * request_values)
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: key_id_len. Data type should be integer.");
     return (1);
   }
-  request_values->key_id = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "key_id_len")->valueint);
+  *key_id = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "key_id_len")->valueint);
   if (!cJSON_IsString(cJSON_GetObjectItem(json, "key_id")))
   {
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: key_id. Data type should be string.");
-    freeCharBuf(&request_values->key_id);
+    freeCharBuf(key_id);
     return (1);
   }
   if (cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring != NULL)
   {
-    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring) != request_values->key_id.len)
+    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring) != key_id->len)
     {
       pelz_log(LOG_ERR, "Length of value in JSON key: key_id does not match value in JSON key: key_id_len.");
-      freeCharBuf(&request_values->key_id);
+      freeCharBuf(key_id);
       return (1);
     }
-    memcpy(request_values->key_id.chars, cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring,
-      request_values->key_id.len);
+    memcpy(key_id->chars, cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring, key_id->len);
   }
   else
   {
     pelz_log(LOG_ERR, "No value in JSON key: key_id.");
-    freeCharBuf(&request_values->key_id);
+    freeCharBuf(key_id);
     return (1);
   }
   if (!cJSON_IsNumber(cJSON_GetObjectItem(json, "enc_data_len")))
   {
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: enc_data_len. Data type should be integer.");
-    freeCharBuf(&request_values->key_id);
+    freeCharBuf(key_id);
     return (1);
   }
-  request_values->data_in = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "enc_data_len")->valueint);
+  *data = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "enc_data_len")->valueint);
   if (!cJSON_IsString(cJSON_GetObjectItem(json, "enc_data")))
   {
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: enc_data. Data type should be string.");
-    freeCharBuf(&request_values->key_id);
-    freeCharBuf(&request_values->data_in);
+    freeCharBuf(key_id);
+    freeCharBuf(data);
     return (1);
   }
   if (cJSON_GetObjectItemCaseSensitive(json, "enc_data")->valuestring != NULL)
   {
-    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "enc_data")->valuestring) != request_values->data_in.len)
+    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "enc_data")->valuestring) != data->len)
     {
       pelz_log(LOG_ERR, "Length of value in JSON key: enc_data does not match value in JSON key: enc_data_len.");
-      freeCharBuf(&request_values->key_id);
-      freeCharBuf(&request_values->data_in);
+      freeCharBuf(key_id);
+      freeCharBuf(data);
       return (1);
     }
-    memcpy(request_values->data_in.chars, cJSON_GetObjectItemCaseSensitive(json, "enc_data")->valuestring,
-      request_values->data_in.len);
+    memcpy(data->chars, cJSON_GetObjectItemCaseSensitive(json, "enc_data")->valuestring, data->len);
   }
   else
   {
     pelz_log(LOG_ERR, "No value in JSON key: enc_data.");
-    freeCharBuf(&request_values->key_id);
-    freeCharBuf(&request_values->data_in);
+    freeCharBuf(key_id);
+    freeCharBuf(data);
     return (1);
   }
   return (0);
 }
 
-int decrypt_parser(cJSON * json, RequestValues * request_values)
+int decrypt_parser(cJSON * json, CharBuf * key_id, CharBuf * data)
 {
   if (!cJSON_HasObjectItem(json, "key_id"))
   {
@@ -246,55 +244,53 @@ int decrypt_parser(cJSON * json, RequestValues * request_values)
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: key_id_len. Data type should be integer.");
     return (1);
   }
-  request_values->key_id = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "key_id_len")->valueint);
+  *key_id = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "key_id_len")->valueint);
   if (!cJSON_IsString(cJSON_GetObjectItem(json, "key_id")))
   {
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: key_id. Data type should be string.");
-    freeCharBuf(&request_values->key_id);
+    freeCharBuf(key_id);
     return (1);
   }
   if (cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring != NULL)
   {
-    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring) != request_values->key_id.len)
+    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring) != key_id->len)
     {
       pelz_log(LOG_ERR, "Length of value in JSON key: key_id does not match value in JSON key: key_id_len.");
-      freeCharBuf(&request_values->key_id);
+      freeCharBuf(key_id);
       return (1);
     }
-    memcpy(request_values->key_id.chars, cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring,
-      request_values->key_id.len);
+    memcpy(key_id->chars, cJSON_GetObjectItemCaseSensitive(json, "key_id")->valuestring, key_id->len);
   }
   else
   {
     pelz_log(LOG_ERR, "No value in JSON key: key_id.");
-    freeCharBuf(&request_values->key_id);
+    freeCharBuf(key_id);
     return (1);
   }
-  request_values->data_in = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "dec_data_len")->valueint);
+  *data = newCharBuf(cJSON_GetObjectItemCaseSensitive(json, "dec_data_len")->valueint);
   if (!cJSON_IsString(cJSON_GetObjectItem(json, "dec_data")))
   {
     pelz_log(LOG_ERR, "Incorrect data type of JSON value of JSON key: dec_data. Data type should be string.");
-    freeCharBuf(&request_values->key_id);
-    freeCharBuf(&request_values->data_in);
+    freeCharBuf(key_id);
+    freeCharBuf(data);
     return (1);
   }
   if (cJSON_GetObjectItemCaseSensitive(json, "dec_data")->valuestring != NULL)
   {
-    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "dec_data")->valuestring) != request_values->data_in.len)
+    if (strlen(cJSON_GetObjectItemCaseSensitive(json, "dec_data")->valuestring) != data->len)
     {
       pelz_log(LOG_ERR, "Length of value in JSON key: dec_data does not match value in JSON key: dec_data_len.");
-      freeCharBuf(&request_values->key_id);
-      freeCharBuf(&request_values->data_in);
+      freeCharBuf(key_id);
+      freeCharBuf(data);
       return (1);
     }
-    memcpy(request_values->data_in.chars, cJSON_GetObjectItemCaseSensitive(json, "dec_data")->valuestring,
-      request_values->data_in.len);
+    memcpy(data->chars, cJSON_GetObjectItemCaseSensitive(json, "dec_data")->valuestring, data->len);
   }
   else
   {
     pelz_log(LOG_ERR, "No value in JSON key: dec_data.");
-    freeCharBuf(&request_values->key_id);
-    freeCharBuf(&request_values->data_in);
+    freeCharBuf(key_id);
+    freeCharBuf(data);
     return (1);
   }
   return (0);
